# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ePortfolio API",
        "description": "API para el sistema de portfolio electrónico de estudiantes",
        "contact": {
            "name": "ePortfolio Support",
            "email": "admin@eportfolio.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local development server"
        },
        {
            "url": "http://eportfolio.test/api/v1",
            "description": "Local development server"
        },
        {
            "url": "https://eportfolio.cifpcarlos3.es/api/v1",
            "description": "Production server"
        },
        {
            "url": "https://eportfolio.juezlti.eu/api/v1",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/familias-profesionales": {
            "get": {
                "tags": [
                    "FamiliaProfesional"
                ],
                "summary": "List all familiaprofesionals",
                "description": "Retrieve a paginated list of familiaprofesionals",
                "operationId": "a566ea7189622f1903ff7b148ec7bd84",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/FamiliaProfesional"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "FamiliaProfesional"
                ],
                "summary": "Create a new familiaprofesional",
                "description": "Create a new familiaprofesional resource",
                "operationId": "c0d9197bc6056d788fd283f870fe2079",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreFamiliaProfesionalRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Resource created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/FamiliaProfesional"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/familias-profesionales/{id}": {
            "get": {
                "tags": [
                    "FamiliaProfesional"
                ],
                "summary": "Show a specific familiaprofesional",
                "description": "Retrieve a specific familiaprofesional by ID",
                "operationId": "0a515eea06013e0640b207178573ae1d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the familiaprofesional",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/FamiliaProfesional"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "FamiliaProfesional"
                ],
                "summary": "Update a specific familiaprofesional",
                "description": "Update a specific familiaprofesional by ID",
                "operationId": "1366b9a9638ec2490e2506567c8fd617",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the familiaprofesional",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateFamiliaProfesionalRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resource updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/FamiliaProfesional"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "FamiliaProfesional"
                ],
                "summary": "Delete a specific familiaprofesional",
                "description": "Delete a specific familiaprofesional by ID",
                "operationId": "8067e4abbe2ef9b52d7e14654e217646",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the familiaprofesional",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Resource deleted successfully"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreFamiliaProfesionalRequest": {
                "title": "Store Familia Profesional Request",
                "description": "Datos requeridos para crear una Familia Profesional",
                "required": [
                    "nombre",
                    "codigo"
                ],
                "properties": {
                    "nombre": {
                        "description": "Nombre de la familia profesional",
                        "type": "string",
                        "maxLength": 255
                    },
                    "codigo": {
                        "description": "Código único de la familia profesional",
                        "type": "string",
                        "maxLength": 50
                    },
                    "descripcion": {
                        "description": "Descripción de la familia profesional",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "UpdateFamiliaProfesionalRequest": {
                "title": "Update Familia Profesional Request",
                "description": "Datos para actualizar una Familia Profesional",
                "properties": {
                    "nombre": {
                        "description": "Nombre de la familia profesional",
                        "type": "string",
                        "maxLength": 255
                    },
                    "codigo": {
                        "description": "Código único de la familia profesional",
                        "type": "string",
                        "maxLength": 50
                    },
                    "descripcion": {
                        "description": "Descripción de la familia profesional",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "FamiliaProfesional": {
                "title": "Familia Profesional",
                "description": "Modelo de Familia Profesional",
                "properties": {
                    "id": {
                        "description": "ID único",
                        "type": "integer"
                    },
                    "nombre": {
                        "description": "Nombre de la familia profesional",
                        "type": "string"
                    },
                    "codigo": {
                        "description": "Código único de la familia profesional",
                        "type": "string"
                    },
                    "descripcion": {
                        "description": "Descripción de la familia profesional",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}